requirements.txt
python-telegram-bot==13.15
openai==1.14.3
python-dotenv==1.0.1
pyyaml==6.0.1
